backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
requests==2.32.3
python-dotenv==1.0.1
gunicorn==22.0.0
redis==5.0.7
PyYAML==6.0.1
Werkzeug==3.0.3
